Create implicit&ExplicitWaits
Implicit , Explicit Wait & Thread.sleep
implicit is defined globally& is applicable at every line of code. It makes the performance slower

Explicit - we wait for a certain condition & then fail the code ( if ) . The scope is limited to a particular element only.

WebDriverWait w =new WebDriverWait(driver,5);		
w.until(ExpectedConditions.visibilityOfElementLocated(By.cssSelector("input.promoCode")));
		

Thread.sleep - not a part of Selenium library, it is a part of Java. It is static condition. Halts the program execution for specified time.